First version of Native App Grant support
This is the initial version of support in the SDK for the Globus Auth
Native App Grant. This grant type will allow applications running
without a web component to complete a full 3-legged OAuth2 flow via the
same mechanisms that other providers have called PIN-based auth.
Basically, a temporary secret and one-way hash of the secret are used to
secure the request for a code, which, when entered, is exchanged for a
token or set of tokens. The code is an auth_code (as from the typical
3-legged OAuth flow), and the tokens are potentially access tokens or
access tokens + refresh tokens.
In particular, we're interested in the possibility of storing refresh
tokens in ~/.globus.cfg in order to request new Access Tokens on an
ongoing basis.
Adds new classes designed to handle OAuth2 flows in general (with Native
App being the only flow thusfar supported), wrappers in the AuthClient
to make access to those classes easier, and a specialized response type
to normalize the token response data from it's spec-compliant format to
a more workable one.
All of the new work is fully-documented in reST-style, so it should be
easy to add to the sphinx site.