Replace deprecated `auth` param with auth disabler
On the basic HTTP methods supported by the BaseClient class, the basic
auth header as specified by the `auth` kwarg is no longer supported. At
the same time, the `no_auth_header` argument was added, as it will be
needed for the AuthClient to POST to Globus Auth with neither client
credentials nor a token (part of the Native App Grant flow).