Fix TransferClient.endpoint_search max results
max_total_results=1000 is incorrect!
The maximum number of *pages* of output is 1000 (offset of 0 to 999) and
the max page size is 100.

max_total_results is measured against the individual results found via
pagination, not the pages. So the correct max_total_results is 1000 *
100 or 100000.

Implementation-wise, `max_total_results` caps the `num_results`
parameter, whereas `max_results_per_call` caps the `limit` parameter
sent to the API.

I'd like to refactor this to be based on `max_offset` or something else
defined in a way more compatible with the API docs, so we don't have
this multiplication buried here. However, I don't think it's reasonable
to go about gutting this when it could have interplay and impact with
other paginated APIs using different paging styles.