Refactor decode_id_token (decompose it)
decode_id_token() previously did two round-trips to Auth + a decoding
step on the JWK prior to handling the ID token. This is prohibitively
expensive if anyone were to try to build a service using
decode_id_token. Instead, you would have to rewrite the function
yourself.

So this is now broken down into two fetch methods, one of which can
optionally do a decoding step, and their results *may* be passed to
decode_id_token. As a result, loading data over the network can always
be taken care of separately from parsing an id token, and parsing a
JWK to a public key object can also be included in those steps.

AuthClient.get_openid_configuration is a new method for fetching
OIDC config.
AuthClient.get_jwk is a new method for getting, and optionally parsing
the JWK.

Unfortunately, the JWK URI provided by OIDC config is a full string
including `https://`, not just the path component. We could strip the
leading `https://auth.globus.org` from it, but I chose to allow
`BaseClient._request` to handle requests to full URLs. It's a single
line of dispatch there vs parsing and unparsing the JWK URI.

get_jwk can take OIDC config as a parameter, or fetch it
automatically.

decode_id_token can take both the OIDC config and the JWK as
paramters, or fetch and use both automatically (and it will pass OIDC
config to the JWK to avoid a double-fetch of that data).

If an application wants to cache the parsed JWK and OIDC config, that
is possible, and they should simply be passed back to the
decode_id_token function. Caching the parsed JWK would require
defining some serialization format, but that's not the SDK's problem
to solve. And we could make things more variable by letting parsing of
the JWK to a public key be separated out even further, but that does
not seem to be necessary.

closes #298