Replace `read_as_dict` with `get_token_data`
This adds a new, guaranteed interface to the token storage adapters.
Although both current adapters implemented `read_as_dict()`, this is
not an interface we can add to the top-level abstract class, as we may
want to add adapters in which enumerating the stored resource servers
is not possible.

Instead, this adds `get_token_data(resource_server)` as a method,
which returns the token information for a given resource server.
Because the adapters we have today *do* make it possible for us to
enumerate resource servers, both implement the
`get_by_resource_server()` method, which produces the correctly
formatted dict.

In order to accommodate these changes, the internal structure and
storage format of the SimpleJSONFileAdapter had to be completely
rewritten. In order to ensure that the new format and storage method
works for handling new tokens (for multiple servers) and a refresh, a
new test is added to the testsuite.

The SimpleJSONFileAdapter now also implements `format_version` and
`supported_versions` class attributes which describe the data format
and what can be supported in terms of loading data. This will give us
leeway to restructure the data as necessary over time, though there is
no known need for this at present. These parameters are part of a
structure which "guarantees" that attempting to load data will produce
the desired format of dict or fail with an error.